requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
google-auth==2.41.1
google-genai==1.45.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMuPDF==1.26.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
